Fix `new` usage in README examples
    x = new require('foo').Bar()
    // is like:
    x = (function () { var x = new require('foo'); return x.Bar()})()

What you really want is something like:

    x = (function () { var x = require('foo').Bar; return new x()})()

so you need parens.

This works by accident, because `new` will return the ctor's
returned object if it's an object, and routes.Router doesn't
actually return an instance anyhow.  However, if it ever did,
and it didn't use an instanceof check, then it would be a bug,
because it would poop on the global object.